fix import by creating *.ts file
