+++ web/src/lib/components/shared-components/settings/setting-input-field.spec.ts
@@ -1,6 +1,7 @@
 import { render } from '@testing-library/svelte';
 import userEvent from '@testing-library/user-event';
-import SettingInputField, { SettingInputFieldType } from './setting-input-field.svelte';
+import { SettingInputFieldType } from './setting-input-field';
+import SettingInputField from './setting-input-field.svelte';
 
 describe('SettingInputField component', () => {
   it('validates number input on blur', async () => {
