+++ web/src/lib/components/shared-components/settings/setting-input-field.ts
@@ -1,0 +1,6 @@
+export enum SettingInputFieldType {
+  EMAIL = 'email',
+  TEXT = 'text',
+  NUMBER = 'number',
+  PASSWORD = 'password',
+}
